+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 1. [Grav](#grav)
     1. [](#improved)
         - Grav Content particle looks now for authors `name`, `alias` or `username` in page header
+        - Added option to cli command `bin/plugin gantry5 child-theme` to clone the theme settings 
     1. [](#bugfix)
         - Fixed outline/particle assignments when `Include default language` in Grav was `No` (#2115)